Added dataset for summer. Made functions.
Choose between summer or winter. Or collect own dataset. Added overwriting to n.snapshots so that it alligns with the snapshot indexes in the datast we aquire from ENTSO. 
Before n.snapshots was equal to the indexes obtained from running snakemake (01.01.2023) now it is overwritten to match with whatever date.